backend: extracts all "controllers" own folder
