+++ packages/backend/src/controllers/index.ts
@@ -1,0 +1,3 @@
+export { generic as renderGeneric, renderSplice } from './renderers';
+export { details as styleDetails, all as allStyles } from './style';
+export { spliceMetadata } from './metadata';
